Start of a channel model for using the connection
This is intended to be a bit like the RabbitMQ Java client's API. This
requires a couple of hooks into the connection, which has to keep
track of channel callbacks and which channels are free.

The channel itself is simply forwarded frames from the connection; it
keeps track of its own outstanding RPCs (and pending methods, since
only one RPC an be outstanding at a time). The frame handler needs to
know which frames are replies, in order to match them with requests,
which is a bit icky. Nonetheless this gives a fairly uniform base on
which to make the API, a few examples of which I've implemented
(`declareQueue`, `publish`).