Support `consume` and declaring/checking exchanges
Also overhauls the frame dispatch into one switch statement, which
makes things a bit more compact (since most kinds of frames fall into
`default`).

One unfortunate side-effect of using `when.js` is that it (in common
with many promise libs so far as I can tell) takes pains to invoke
callbacks outside the thread of control; that is, it uses `nextTick`
or equivalent to run them once a value has been supplied. This creates
a race for me, when registering consumers: I need the consumerTag
which is in the reply, but if my callback is called nextTick, messages
may have arrived in the meantime and have nowhere to go.

To avoid that I've made a dirt simple promise lib that always invokes
its callbacks in the current thread of control, and used that for
RPCs; that way I know any effects will be complete before more frames
are processed.