Start tests for connection
Involves a bit of machinery to be able to build protocol exchanges to
look like a server. In the end I used more or less the same machinery
that the connection uses itself (composing promises).

Interesting bug this threw up: the original implementation of the
opening handshake used a chain of promises, with the continuation for
certain steps setting the `.accept` method (called for each frame
received) to resolve the next promise in the sequence. Once the chain
of responses was in place, the connection was set to run, that is
process frames. However `run` processes all the frames available at a
time, and my test machinery effectively does synchronous sends, which
in combination meant that a frame received by the mock server could
end up invoking the connection's `accept` method *before* the promise
continuation, typically scheduled with `nextTick`, had a chance to run
(and thereby replace the accept method).

One way to avoid this would be for the promise continuations to be
invoked synchronously, so it all proceeded in lock step; and I may yet
do that anyway. But my solution was to make a `step` method, which
tells the connection to receive a single frame when one is
available. This is a bit more explicit than using synchronous promises
(and may be easier to test, and assert invariants, for that reason).