Implement BasicReturn, -Ack and -Nack
BasicReturn is a 'content bearing' method, so it can use the same
state machine as BasicDeliver. It turns out I had that state machine
wrong, as well: the starting state should be 'accept a deliver or
return', not 'I don't know what this is'. As it was, an interleaved
deliver or return would effectively reset the state.