Try to provide a useful error if the channel closes on an RPC
If the server closes the channel, we may get a ChannelClose instead of
an RPC reply; since there's only one outstanding RPC at a time, it
will either be because of that RPC, or something asynchronous (a bad
publish maybe, or a human forcing a close). So, we can check the
stated reason and give a better error message if it was indeed the
RPC.

Also: fixes a daft bug with rejecting pending RPCs and provides a test
for it; actually provide values for `closedSend` and `closedAccept` in
connection shutdown.