Make the promise-chaining a bit more bearable
1. provide `doAll`, which assures all operations pass (not the case
before; operations done sequentially could hide failures)
2. compact the `then`ning where possible, which I find more readable

On that second: there's often a choice between immediately returning a
promise, thereby staying at the same 'indent level'; and appending a
`then` to the return value. I've generally used the latter when the
extra operation will fit in one line (in the tests themselves); or it
needs to refer to a variable in its closure (in the scaffolding
usually).