Negotiate tuning during connection open
Sounds more glamorous than it is. Essentially we get some values from
the server for channelMax, frameMax(Size) and heartbeat (interval),
accept or lower them or indicate no such limit (or interval), then the
server decides whether to enforce what we've chosen.

The channel_api tests, that speak to an actual server, are
parameterised with the environment variable `URL` now, so different
tuning parameters can be tried with e.g.,

    URL='amqp://?frameMax=0x1000&channelMax=6&heartbeat=300' make test

Once it was possible to turn heartbeating on it became clear heartbeat
frames weren't really handled; I've fixed that by making the heartbeat
frame a constant and having a case for it in "channel0".