Expand README and add client API reference docs
This removes much of the commentary from channel_api.js and api.js, in
favour of giving it as documentation external to the code. My decision
to do this was influenced by none of the inline reference
documentation tools being especially good; in particular I couldn't
get jsdoc to work acceptably at all.

Mostly though, I prefer to have explanation of the API, rather than
the code, outside the code.