Distinguish among errors thrown while processing messages
There's three sources of errors: the server screwing up, the library
screwing up, or the application screwing up. To date, errors thrown
while handling messages were not accounted for; worse, they upset the
machinery because it uses exceptions to signal server errors.

All the varieties of error have the same ultimate consequence; however
it's important to separate our own error handling from what we tell
the server. For this reason I've separated `reason` (the message for
the server) and `error` (the actual error reported in the client),
rather than constructing the error from the reason.