Avoid stream gotchas in prior versions of Node/readable-stream
`readable-stream` v1.0.2 would not fire `'readable'` if the listener
was registered after writes, usually necessitating an explicit
invocation of the handler (in this case, `enqueue` in
`Mux#pipeFrom`). This was fixed upstream in Node.JS v0.10.something
and backported to `readable-stream` at some point before v1.1.9 (the
most recent at the time of commit).

Node.JS v0.9 does not need `readable-stream`, but doesn't appear to
have the backported fixes from v0.10. For the sake of completeness,
I've rejigged the mux roundrobin test to avoid that particular race.

`readable-stream` also monkey-patches `setImmediate` into the global
object if it's not present. However, I don't want to rely on that so
I've used my own, local solution.