Use callbacks instead of promises for ConfirmChannel#publish
A disadvantage to returning promises from ConfirmChannel#publish (and
thereby ConfirmChannel#sendToQueue) is that it can no longer return a
boolean indicating write buffer overflow. So instead of returning a
promise, take a callback as the last argument. It's easy enough to
make it resolve a promise you've cons'ed beforehand.