Advance the unconfirmed window over gaps
If acks arrive out of order, there may be gaps in the window. Whenever
we would advance the low water mark -- that is, when the tag is the
same as the low water mark -- we have to keep moving the window until
its either empty, or there's an unconfirmed message. To be able to
distinguish between entries that have already been resolved and
entries that are undefined (e.g., because a callback argument was not
supplied), `null` is used to mark resolved entries, and falsey
callbacks (which might otherwise be null) are replaced with `false`
specifically.