Implement consumer cancel notification
This is a RabbitMQ extension: if a consumer will not receive any more
messages, RabbitMQ will send a cancel frame.

The most straight-forward, and I think idiomatic, way of representing
this is by calling the consumer callback with `null` instead of a
message. This lines up with how readable streams work (push `null` to
indicate no more values), and doesn't require any changes to the API
(other than the possibility of getting called with a null).