Write message frames all at once if small enough
If the method, properties and content frames all fit in 2KB, copy them
into a single buffer and write that. The 2KB threshold is more or less
arbitrary, same as the "safely-sized" buffers allocated when encoding
-- it's based on some fairly vague reasoning around being sufficiently
smaller than the SlowBuffer allocation size.  The method and
properties for a message is on the order of 100 bytes (depending on
the strings supplied), so there's room for almost the full 2KB of
message content.

It turns out this is quite a boost for those small messages, even
though it involves more buffer allocations. Socket writes are just
expensive!

This further coalesces `sendMethodAndProperties`+`sendContent` to
`sendMessage`. Not sure why I had them separate in the first place.