Allocate buffers after the encoded size is known
Instead of checking whether every field is fixed and if not,
allocating a large buffer, calculate the size of fields that aren't
fixed-size first, then allocate the buffer to an exact size.

In the case of strings this just means getting the byte length
(assuming they are UTF8 as the AMQP specification requires). In the
case of tables, I have to encode them into a buffer first to get the
size; and once I've done that, I can keep the encoded table around to
copy into the full result.

Properties are a bit trickier because of the flags indicating
presence; this makes every field effectively variable size, since it
may or may not be present. I end up with two passes, one to calculate
the size and one to write the values.