Declare the decoding result with all fields
As I understand it, this avoids V8 switching the object's hidden class
every time a value is assigned to a field. The trade off is that V8
might allocate larger object than it needs to -- especially in the
case of properties, which may quite reasonably be empty. For that
reason, there's a shortcut when no flags are set in an encoded
properties.