Don't emit redundant (and invalid) variable ref
The code generation had a format call that provided one too many
arguments, resulting in a trailing variable reference that hadn't been
defined. This only shows up if the file is not uglified, since that
will remove the redundant expression.

Relatedly, it was possible to generate the defs.js file without then
uglifying it, by running `make` before `npm install`.