Remove test for RabbitMQ behaviour which has changed
According to the AMQP spec, trying to delete a queue or exchange that
doesn't exist is an error. At v3.2.0, RabbitMQ's behaviour is to *not*
treat it as an error (it just trivially succeeds). Since the test was
testing the server semantics rather than the client behaviour, I have
no problem removing it.