Rewrite frame-coalescing for three cases
In order of presumed likelihood:

 1. there's a short message and all of the method, properties, and
 content fit in a single chunk
 2. there's a longer message, but the method and headers fit in a
 single chunk
 3. all of the method, headers, and content must be sent in separate
 chunks

There's no attempt to account for the other case, that is the
properties and content fit in a single chunk but the method must be
sent on its own (first). I'm not sure that could ever be the case,
since there's limits to the field sizes of BasicPublish, though of
course that depends on the arbitrary threshold for what is a "single
chunk".

While testing this I noticed that a large value in the headers table
can make it bigger than the encoding `SCRATCH` area; this results in
an OOB exception in Node.JS 0.8, and silently succeeds (!) in Node.JS
> 0.8. I've increased the size of the scratch are to at least be
double the single chunk threshold (so it's in a sense consistent, if
still fixed and arbitrary). Not sure what to do about that in the
longer term.