Fail properly in APIs when no channel number available
Previously, the two APIs would throw an exception when a channel
number couldn't be allocated, instead of failing in the idiomatic way
(rejecting the promise or calling the callback with an error in the
first argument).

To make this work, I had to separate the allocation of the number from
the RPC, and do both steps in each API implementation. Another way
would be to give freshChannel a callback argument.