Emit error with close, or none if client-initiated
There are four ways a connection closes:

 * if the server decides the connection is in error (e.g., bad frame)
 * the client decides the connection is in error (e.g., socket
   malfunction)
 * if an operator closes the connection, or shuts down the server
 * if the application code closes the connection

In general, we want to know if it's worth crashing. If the application
code closes the connection, it's not worth crashing.  If the server or
client decided the connection was broken, it's worth crashing. If the
operator decided to close the connection or the server is restarting,
it might be worth crashing.

Crashing is what the 'error' event is for. So: if it's unambiguously
bad (first two cases), there's an 'error' event. In the last case, I
let the applicaton decide, by only emiting 'close', but supplying the
error (and a predicate to say whether it was a bad
error).