Clear up credentials handling.
After revisiting the specification
https://www.rabbitmq.com/uri-spec.html and thus
http://www.ietf.org/rfc/rfc3986.txt, I've made credentials handling
more consistent for both stringly URLs and object URLs.

 - iff both parts are missing, use defaults
 - if anything is supplied:
   - the username ends before the first colon
   - the password starts after the first colon
   - absent values are treated as empty

On that last: the spec says "do what you want with absent values", and
another option is to replace absent values with a default. But I think
it makes more sense to treat username-and-password as a unit you
supply all or none of.

This commit adds some tests for credentials parsing, and makes the
parsing more explicit (if a bit longer).