Use a third-party URL parser
The `url.parse` from the standard lib provides the username and
password as one field, and decodes them. This means any
percent-encoded characters are decoded, and it's impossible to tell
whether a colon is the delimiter between the username and password, or
was encoded and part of one or other.

Newer versions of the Node standard library come with an updated URL
parser. However, I would like to work with older versions of Node, so
I've used a third party parser.