Make sure channel buffers empty when closing
When a channel is closed, its buffer is unpiped from the connection
muxer as part of cleaning up. In NodeJS v10, this happens before any
frames that have just been written to the channel are picked up by the
muxer -- e.g., the CloseOK response.

Instead of directly unpiping, end() the channel buffer, and let it be
unpiped once all its frames are read. So far as I can think it
through, this is correct behaviour anyway, and it only worked before
by coincidence.