Bump mocha to fix stray failures
The test suite fails on a couple of test cases which succeed
individually, just in older versions of Node. Running the tests with a
path ending in test/*.js (rather than test/) averts the test
failures. So does upgrading mocha.

I have bumped it to 3.5.3, which is the newest version that still
supports NodeJS <4.

While I was at it, I upgraded a few other packages.