Introduce post-connection timeout setting
Currently, the timeout for the underlying socket is reset to `0` after a successful connection. This can lead to promises hanging indefinitely / callback not being invoked in case the server become unresponsive. On the other hand, just using the connection timeout may be too short, and introduce runtime errors.

As a solution, introduce a second socket option `postConnectTimeout` which is being applied to the socket rather than resetting the timeout to `0`. If the setting is not provided, a high default timeout of 300000 (5 minutes) is being set, which would still prevent indefinitely blocking clients.